Sessions management

+++ backend/index.js
@@ -73,6 +73,7 @@ app.post("/api/login", (req, res) => {
               if(result){
                   // Start session
                   req.session.usuario=user;
+                  req.session.userType = result.userType;
                   req.session.nulidad = result.nulidad;
                   req.session.investigacion = result.investigacion;
                   // Send session
@@ -101,42 +102,45 @@ app.post("/api/register", (req, res) => {
   let uType = req.body.userType;
   let typeNulidad = req.body.nulidad; 
   let typeInv = req.body.investigacion;
-  try{
-    // Look for user in database
-    db.collection("usuarios").findOne({email:mail}, (err, result) => {
-      // If user exists
-      if(result!=null)
-      {
-        res.send({"Message": "User already exists"});
-      }
-      // Check if password confirmation is valid
-      else if (pass != cpass) 
-      {
-        res.send({"Message": "Passwords don't match"});
-      }
-      // Otherwise
-      else
-      {
-        // Hash password
-        bcrypt.hash(pass, 10, (err, hash) => {
-          // Save new user info into variable
-          let aAgregar = {usuario: user, email: mail, password: hash, userType: uType, nulidad: typeNulidad, investigacion: typeInv}
-          // Add new user info to database
-          db.collection("usuarios").insertOne(aAgregar, (err, result) => {
-          // Evaluate for error
-          if(err) throw err;
-          res.send({"Message": "User registered"});
+  if(req.session.usuario && req.session.userType == "Administrador")
+  {
+    try{
+      // Look for user in database
+      db.collection("usuarios").findOne({email:mail}, (err, result) => {
+        // If user exists
+        if(result!=null)
+        {
+          res.send({"Message": "User already exists"});
+        }
+        // Check if password confirmation is valid
+        else if (pass != cpass) 
+        {
+          res.send({"Message": "Passwords don't match"});
+        }
+        // Otherwise
+        else
+        {
+          // Hash password
+          bcrypt.hash(pass, 10, (err, hash) => {
+            // Save new user info into variable
+            let aAgregar = {usuario: user, email: mail, password: hash, userType: uType, nulidad: typeNulidad, investigacion: typeInv}
+            // Add new user info to database
+            db.collection("usuarios").insertOne(aAgregar, (err, result) => {
+            // Evaluate for error
+            if(err) throw err;
+            res.send({"Message": "User registered"});
+            })
           })
-        })
-      }
-    })
-    res.json({'Message': "User inserted correctly."});
-  }
-  // Error
-  catch (error){
-    res.status(500);
-    res.json(error);
-    console.log(error);
+        }
+      })
+      res.json({'Message': "User inserted correctly."});
+    }
+    // Error
+    catch (error){
+      res.status(500);
+      res.json(error);
+      console.log(error);
+    }
   }
 })
 
@@ -183,26 +187,32 @@ app.get("/api/sessionExists", async (req, res) => {
 
 // Get for all users in database
 app.get("/api/getAllUsers", async (req, res) => {
-  try {
-    // Get all info from users collection
-    const cursor = db.collection("usuarios").find(); 
-    const data = await cursor.toArray();
-    // Get results
-    res.json(data);
-  } 
-  // Error
-  catch (error) {
-    res.status(500);
-    res.json(error);
-    console.log(error);
+  if(req.session.usuario)
+  {
+    try {
+      // Get all info from users collection
+      const cursor = db.collection("usuarios").find(); 
+      const data = await cursor.toArray();
+      // Get results
+      res.json(data);
+    } 
+    // Error
+    catch (error) {
+      res.status(500);
+      res.json(error);
+      console.log(error);
+    }
   }
 })
 
 // Get for logout
 app.get("/api/logout", (req, res) => {
-  // Delete session to logout
-  req.session.destroy();
-  res.json({});
+  if(req.session.usuario)
+  {
+    // Delete session to logout
+    req.session.destroy();
+    res.json({});
+  }
 })
 
 // Delete for user
@@ -210,9 +220,12 @@ app.delete("/api/deleteUser", (req, res) => {
   // Get req user data
   let mail = req.body.email;
   let user = req.body.usuario;
-  // Delete from database
-  db.collection("usuarios").deleteOne({usuario: user, email: mail});
-  res.send({"message": "User deleted"});
+  if(req.session.usuario && req.session.userType == "Administrador")
+  {
+    // Delete from database
+    db.collection("usuarios").deleteOne({usuario: user, email: mail});
+    res.send({"message": "User deleted"});
+  }
 })
 
 // Put to edit users
@@ -223,9 +236,12 @@ app.put("/api/editUser", (req, res) => {
   let ogEmail = req.body.ogEmail;
   let tNulidad = req.body.nulidad; 
   let tInvestigacion = req.body.investigacion;
-  // Update database
-  db.collection("usuarios").updateOne({email: ogEmail}, {$set: {usuario: user, email: mail, nulidad: tNulidad, investigacion: tInvestigacion}});
-  res.send({"Message": "User edited"});
+  if(req.session.usuario && req.session.userType == "Administrador")
+  {
+    // Update database
+    db.collection("usuarios").updateOne({email: ogEmail}, {$set: {usuario: user, email: mail, nulidad: tNulidad, investigacion: tInvestigacion}});
+    res.send({"Message": "User edited"});
+  }
 })
 
 // Put to update passwords
@@ -235,45 +251,48 @@ app.put("/api/changePass", (req, res) => {
   let ogPass = req.body.ogPassword;
   let newPass = req.body.newPassword;
   let repPass = req.body.repPassword;
-  // Look for user in database
-  db.collection("usuarios").findOne({email:mail}, (err, result) => {
-    // If user exists
-    if(result!=null)
-    {
-      // Verify if original password matches password in database
-      bcrypt.compare(ogPass, result.password, (err, result) => {
-        // If they match
-        if(result){
-          // Confirm new password
-          if (newPass != repPass) 
-          {
-            // If new passwords don't match
-            res.send({"Message": "Passwords don't match"})
+  if(req.session.usuario && req.session.userType == "Administrador")
+  {
+    // Look for user in database
+    db.collection("usuarios").findOne({email:mail}, (err, result) => {
+      // If user exists
+      if(result!=null)
+      {
+        // Verify if original password matches password in database
+        bcrypt.compare(ogPass, result.password, (err, result) => {
+          // If they match
+          if(result){
+            // Confirm new password
+            if (newPass != repPass) 
+            {
+              // If new passwords don't match
+              res.send({"Message": "Passwords don't match"})
+            }
+            // If new passwords match
+            else {
+              // Hash new password
+              bcrypt.hash(newPass, 10, (err, hash) => {
+                // Error
+                if (err) throw err;
+                // update password in database
+                db.collection("usuarios").updateOne({email: mail}, {$set: {password: hash}})
+              })
+              res.send({"Message": "Change Successful"})
+            }
           }
-          // If new passwords match
-          else {
-            // Hash new password
-            bcrypt.hash(newPass, 10, (err, hash) => {
-              // Error
-              if (err) throw err;
-              // update password in database
-              db.collection("usuarios").updateOne({email: mail}, {$set: {password: hash}})
-            })
-            res.send({"Message": "Change Successful"})
+          // If original password doesn't match with database
+          else{
+            res.send({"Message": "Error"})
           }
-        }
-        // If original password doesn't match with database
-        else{
-          res.send({"Message": "Error"})
-        }
-      })
-    }
-    // If user doesn't exist
-    else
-    {
-      res.send(false)
-    }
-})
+        })
+      }
+      // If user doesn't exist
+      else
+      {
+        res.send(false)
+      }
+    })
+  }
 })
 
 // Post to setup app backend --> create roles 
@@ -311,110 +330,116 @@ app.post("/api/setup", (req, res)=>{
 
 // Upload file path --> Expediente
 app.post("/api/addpath", uploads.single("file"), (req, res) => {
-  try {
-    // Get request data
-    let body = req.body;
-    let collection = req.body.docType;
-    delete(body.folio)
-    delete(body.docType)
-    delete(body.nombre)
-
-    // Insert into collection (nulidad or investigacion)
-    db.collection(collection).insertOne(body, (err,res) => {
-      if (err) throw err;
-      console.log("Expediente Guardado");
-    })
-    res.json({'message': "Data inserted correctly."});
-    // ERror
-  } catch (error) {
-    res.status(500);
-    res.json(error);
-    console.log(error);
+  if(req.session.usuario)
+  {
+    try {
+      // Get request data
+      let body = req.body;
+      let collection = req.body.docType;
+      delete(body.folio)
+      delete(body.docType)
+      delete(body.nombre)
+  
+      // Insert into collection (nulidad or investigacion)
+      db.collection(collection).insertOne(body, (err,res) => {
+        if (err) throw err;
+        console.log("Expediente Guardado");
+      })
+      res.json({'message': "Data inserted correctly."});
+      // ERror
+    } catch (error) {
+      res.status(500);
+      res.json(error);
+      console.log(error);
+    }
   }
 })
 
 // Post for first folio
-app.post("/api/addFirstFolio", uploads.single("file"), async (req,res) => {
-  try {
-    // Get req data
-    let collection = req.body.docType;
-    console.log(collection);
-    // Get key and initialization vector depending on docType
-    if(collection == "juicioNulidad")
-    {
-      console.log("estoy en nulidad")
-      // Nulidad iv and key
-      db.collection("roles").findOne({rol:"nulidad"}, (err, result)=>{
-        console.log("estoy en colección");
-        fs.readFile("testLab.key", (err, decryptKey)=>{
-            console.log("el resultado",result.iv);
-            let key=Buffer.from(crypto.privateDecrypt(decryptKey, Buffer.from(result.llave, "hex")));
-            let iv=Buffer.from(crypto.privateDecrypt(decryptKey, Buffer.from(result.iv, "hex")));
-            uploadFirstFolio(req, key,iv);
+app.post("/api/addFirstFolio", uploads.single("file"), (req,res) => {
+  if(req.session.usuario)
+  {
+    try {
+      // Get req data
+      let collection = req.body.docType;
+      console.log(collection);
+      // Get key and initialization vector depending on docType
+      if(collection == "juicioNulidad")
+      {
+        // Nulidad iv and key
+        db.collection("roles").findOne({rol:"nulidad"}, (err, result)=>{
+          fs.readFile("testLab.key", (err, decryptKey)=>{
+              console.log("el resultado",result.iv);
+              let key=Buffer.from(crypto.privateDecrypt(decryptKey, Buffer.from(result.llave, "hex")));
+              let iv=Buffer.from(crypto.privateDecrypt(decryptKey, Buffer.from(result.iv, "hex")));
+              uploadFirstFolio(req, key,iv);
+          })
         })
-      })
-    }
-    else if (collection == "carpetaInvestigacion")
-    {
-      // Investigacion iv and key
-      db.collection("roles").findOne({rol:"investigacion"}, (err, result)=>{
-        console.log("el resultado",result.iv);
-        fs.readFile("testLab.key", (err, decryptKey)=>{
-            console.log("wuwuwuw");
-            let key=Buffer.from(crypto.privateDecrypt(decryptKey, Buffer.from(result.llave, "hex")));
-            let iv=Buffer.from(crypto.privateDecrypt(decryptKey, Buffer.from(result.iv, "hex")));
-            console.log("el iv", iv)
-            uploadFirstFolio(req, key,iv);
+      }
+      else if (collection == "carpetaInvestigacion")
+      {
+        // Investigacion iv and key
+        db.collection("roles").findOne({rol:"investigacion"}, (err, result)=>{
+          fs.readFile("testLab.key", (err, decryptKey)=>{
+              console.log("wuwuwuw");
+              let key=Buffer.from(crypto.privateDecrypt(decryptKey, Buffer.from(result.llave, "hex")));
+              let iv=Buffer.from(crypto.privateDecrypt(decryptKey, Buffer.from(result.iv, "hex")));
+              console.log("el iv", iv)
+              uploadFirstFolio(req, key,iv);
+          })
         })
-      })
+      }
+      res.json({'message': "Data inserted correctly."});
+      // Error
+    } catch (error) {
+      res.status(500);
+      res.json(error);
+      console.log(error);
     }
-    res.json({'message': "Data inserted correctly."});
-    // Error
-  } catch (error) {
-    res.status(500);
-    res.json(error);
-    console.log(error);
   }
 })
 
 // Put to add folios
 app.put("/api/addfolio", uploads.single("fileFolio"), (req, res) => {
-  try {
-    // Get folio type
-    let collection = req.body.docType;
-    // Get key and initialization vector depending on docType
-    if(collection == "juicioNulidad")
-    {
-      // Nulidad iv and key
-      db.collection("roles").findOne({rol:"nulidad"}, (err, result)=>{
-        fs.readFile("testLab.key", (err, decryptKey)=>{
-            let key=Buffer.from(crypto.privateDecrypt(decryptKey, Buffer.from(result.llave, "hex")));
-            let iv=Buffer.from(crypto.privateDecrypt(decryptKey, Buffer.from(result.iv, "hex")));
-            uploadFolio(req, key, iv);
+  if(req.session.usuario)
+  {
+    try {
+      // Get folio type
+      let collection = req.body.docType;
+      // Get key and initialization vector depending on docType
+      if(collection == "juicioNulidad")
+      {
+        // Nulidad iv and key
+        db.collection("roles").findOne({rol:"nulidad"}, (err, result)=>{
+          fs.readFile("testLab.key", (err, decryptKey)=>{
+              let key=Buffer.from(crypto.privateDecrypt(decryptKey, Buffer.from(result.llave, "hex")));
+              let iv=Buffer.from(crypto.privateDecrypt(decryptKey, Buffer.from(result.iv, "hex")));
+              uploadFolio(req, key, iv);
+          })
         })
-      })
-    }
-    else if (collection == "carpetaInvestigacion")
-    {
-      // Investigacion iv and key
-      db.collection("roles").findOne({rol:"investigacion"}, (err, result)=>{
-        fs.readFile("testLab.key", (err, decryptKey)=>{
-            let key=Buffer.from(crypto.privateDecrypt(decryptKey, Buffer.from(result.llave, "hex")));
-            let iv=Buffer.from(crypto.privateDecrypt(decryptKey, Buffer.from(result.iv, "hex")));
-            uploadFolio(req, key, iv);
+      }
+      else if (collection == "carpetaInvestigacion")
+      {
+        // Investigacion iv and key
+        db.collection("roles").findOne({rol:"investigacion"}, (err, result)=>{
+          fs.readFile("testLab.key", (err, decryptKey)=>{
+              let key=Buffer.from(crypto.privateDecrypt(decryptKey, Buffer.from(result.llave, "hex")));
+              let iv=Buffer.from(crypto.privateDecrypt(decryptKey, Buffer.from(result.iv, "hex")));
+              uploadFolio(req, key, iv);
+          })
         })
-      })
+      }
+      res.json({'message': "Data updated correctly"})
+      // Error
+    } catch (error) {
+      res.status(500);
+      res.json(error);
+      console.log(error);
     }
-    res.json({'message': "Data updated correctly"})
-    // Error
-  } catch (error) {
-    res.status(500);
-    res.json(error);
-    console.log(error);
   }
 })
 
-// Function that 
+// Function that uploads first folio
 async function uploadFirstFolio(req, key, iv)
 {
     // Get req data
@@ -451,6 +476,7 @@ async function uploadFirstFolio(req, key, iv)
     })
 }
 
+// Function to upload folios
 async function uploadFolio(req, key, iv)
 {
   // Define route, input and output
@@ -478,29 +504,34 @@ async function uploadFolio(req, key, iv)
 
 // Post to get docs
 app.post("/api/getDocs", async (request, response) => {
-  try {
-    // Query to search docs
-    let searchValue = {}
-    if (request.body.query != null) {
-      searchValue = request.body.query
+  if(request.session.usuario)
+  {
+    try {
+      // Query to search docs
+      let searchValue = {}
+      if (request.body.query != null) {
+        searchValue = request.body.query
+      }
+      // {"docID" : {$regex : request.body.docID}
+      // Look in collection corresponding to document type (nulidad/investigacion)
+      const cursor = db.collection(request.body.docType).find(searchValue, request.body.projection);
+      const data = await cursor.toArray();
+      console.log(data);
+      // Get data
+      response.json(data);
+      // Error
+    } catch (error) {
+      response.status(500);
+      response.json(error);
+      console.log(error);
     }
-    // {"docID" : {$regex : request.body.docID}
-    // Look in collection corresponding to document type (nulidad/investigacion)
-    const cursor = db.collection(request.body.docType).find(searchValue, request.body.projection);
-    const data = await cursor.toArray();
-    console.log(data);
-    // Get data
-    response.json(data);
-    // Error
-  } catch (error) {
-    response.status(500);
-    response.json(error);
-    console.log(error);
   }
 })
 
 // Post to get folios
 app.post("/api/getFolios", async (request, response) => {
+  if(req.session.usuario)
+  {}
   try {
     // Save query
     let searchValue = request.body.query
@@ -548,36 +579,38 @@ async function descargarArchivo(req, res, key, iv){
 
 // Post to download folio
 app.post("/api/descargarFolio", (req, res) => {
-  console.log(req.body);
   // Initialize key and iv, get doctype to determine them
   let key=""
   let iv= ""
   let folioType = req.body.docType;
-  // Get key and initialization vector depending on docType
-  if(folioType == "juicioNulidad")
+  if(req.session.usuario)
   {
-    // Nulidad iv and key
-    db.collection("roles").findOne({rol:"nulidad"}, (err, result)=>{
-      fs.readFile("testLab.key", (err, decryptKey)=>{
-          key=Buffer.from(crypto.privateDecrypt(decryptKey, Buffer.from(result.llave, "hex")));
-          iv=Buffer.from(crypto.privateDecrypt(decryptKey, Buffer.from(result.iv, "hex")));
-           // Call download function
-          descargarArchivo(req, res, key, iv);
-        })
-      })
-    }
-    else if (folioType == "carpetaInvestigacion")
+    // Get key and initialization vector depending on docType
+    if(folioType == "juicioNulidad")
     {
-      // Investigacion iv and key
-      db.collection("roles").findOne({rol:"investigacion"}, (err, result)=>{
+      // Nulidad iv and key
+      db.collection("roles").findOne({rol:"nulidad"}, (err, result)=>{
         fs.readFile("testLab.key", (err, decryptKey)=>{
             key=Buffer.from(crypto.privateDecrypt(decryptKey, Buffer.from(result.llave, "hex")));
             iv=Buffer.from(crypto.privateDecrypt(decryptKey, Buffer.from(result.iv, "hex")));
-            // Call download function
+             // Call download function
             descargarArchivo(req, res, key, iv);
+          })
         })
-      })
-    }
+      }
+      else if (folioType == "carpetaInvestigacion")
+      {
+        // Investigacion iv and key
+        db.collection("roles").findOne({rol:"investigacion"}, (err, result)=>{
+          fs.readFile("testLab.key", (err, decryptKey)=>{
+              key=Buffer.from(crypto.privateDecrypt(decryptKey, Buffer.from(result.llave, "hex")));
+              iv=Buffer.from(crypto.privateDecrypt(decryptKey, Buffer.from(result.iv, "hex")));
+              // Call download function
+              descargarArchivo(req, res, key, iv);
+          })
+        })
+      }
+  }
 })
 
 // Https server
